services/asl-model/requirements.txt
# ASL Fingerspelling Recognition System
# Requirements file

# Core ML
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0

# MediaPipe for hand landmark detection
mediapipe>=0.10.9

# Computer Vision
opencv-python>=4.8.0
pillow>=10.0.0

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0

# Training & Evaluation
tensorboard>=2.15.0
scikit-learn>=1.3.0
tqdm>=4.66.0
matplotlib>=3.7.0
seaborn>=0.12.0
